замена типа данных у поля id ResultQuizDt.kt

+++ app/src/main/java/ru/vsls/surfquiz/data/local/QuizLocalRepositoryImpl.kt
@@ -15,7 +15,7 @@ class QuizLocalRepositoryImpl @Inject constructor(private val api: QuizDao): Qui
         return api.getAllResults()
     }
 
-    override suspend fun deleteById(id: Int) {
+    override suspend fun deleteById(id: Long) {
         api.deleteById(id)
     }
 }
