+++ app/src/main/java/ru/vsls/surfquiz/domain/repository/QuizLocalRepository.kt
@@ -5,5 +5,5 @@ import ru.vsls.surfquiz.data.local.entities.ResultQuizDt
 interface QuizLocalRepository {
         suspend fun saveResult(result: ResultQuizDt)
         suspend fun getHistory(): List<ResultQuizDt>
-        suspend fun deleteById(id: Int)
+        suspend fun deleteById(id: Long)
 }
