+++ app/src/main/java/ru/vsls/surfquiz/domain/usecase/DeleteQuizHistoryEntryUseCase.kt
@@ -4,11 +4,11 @@ import ru.vsls.surfquiz.domain.repository.QuizLocalRepository
 import javax.inject.Inject
 
 interface DeleteQuizHistoryEntryUseCase {
-    suspend operator fun invoke(id: Int)
+    suspend operator fun invoke(id: Long)
 }
 
 class DeleteQuizHistoryEntryUseCaseImpl @Inject constructor(
     private val repository: QuizLocalRepository,
 ) : DeleteQuizHistoryEntryUseCase {
-    override suspend fun invoke(id: Int) = repository.deleteById(id)
+    override suspend fun invoke(id: Long) = repository.deleteById(id)
 }
